Refactor command handling and enhance error reporting

+++ Command/DimensionCommand.h
@@ -12,8 +12,8 @@
 namespace dev::command {
 
 struct Dimension {
-    int width;
-    int height;
+    uint16_t width;
+    uint16_t height;
 };
 
 class DimensionCommand : public AbstractCommand {
